requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
frozenlist==1.4.0
multidict==6.0.4
yarl==1.9.3